+++ packages/admin-ui/src/lib/core/src/shared/pipes/locale-currency-name.pipe.ts
@@ -32,7 +32,7 @@ export class LocaleCurrencyNamePipe extends LocaleBasePipe implements PipeTransf
         }
         let name = '';
         let symbol = '';
-        const activeLocale = typeof locale === 'string' ? locale : this.locale ?? 'en';
+        const activeLocale = this.getActiveLocale(locale);
 
         // Awaiting TS types for this API: https://github.com/microsoft/TypeScript/pull/44022/files
         const DisplayNames = (Intl as any).DisplayNames;