+++ packages/core/e2e/fixtures/test-plugins/hydration-test-plugin.ts
@@ -6,9 +6,11 @@ import {
     Ctx,
     EntityHydrator,
     ID,
+    LanguageCode,
     OrderService,
     PluginCommonModule,
     Product,
+    ProductService,
     ProductVariantService,
     RequestContext,
     TransactionalConnection,
@@ -23,6 +25,7 @@ export class TestAdminPluginResolver {
         private orderService: OrderService,
         private channelService: ChannelService,
         private productVariantService: ProductVariantService,
+        private productService: ProductService,
         private entityHydrator: EntityHydrator,
     ) {}
 
@@ -65,6 +68,26 @@ export class TestAdminPluginResolver {
             relations: ['product.facetValues.facet'],
         });
         return variant;
+    }
+
+    // Test case for https://github.com/vendure-ecommerce/vendure/issues/1324
+    @Query()
+    async hydrateProductWithNoFacets(@Ctx() ctx: RequestContext) {
+        const product = await this.productService.create(ctx, {
+            enabled: true,
+            translations: [
+                {
+                    languageCode: LanguageCode.en,
+                    name: 'test',
+                    slug: 'test',
+                    description: 'test',
+                },
+            ],
+        });
+        await this.entityHydrator.hydrate(ctx, product, {
+            relations: ['facetValues', 'facetValues.facet'],
+        });
+        return product;
     }
 
     // Test case for https://github.com/vendure-ecommerce/vendure/issues/1172
@@ -110,6 +133,7 @@ export class TestAdminPluginResolver {
         schema: gql`
             extend type Query {
                 hydrateProduct(id: ID!): JSON
+                hydrateProductWithNoFacets: JSON
                 hydrateProductAsset(id: ID!): JSON
                 hydrateProductVariant(id: ID!): JSON
                 hydrateOrder(id: ID!): JSON